pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-wtf>=1.2.2",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "markdown>=3.9",
    "markupsafe>=3.0.3",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "qrcode>=8.2",
    "flask-limiter>=4.0.0",
]
